restructure v1::sinks and v1::sinks::kafka to better encapsulate storage-specific impls
